Added Mr. Ms. and Mrs. to hello.py
Added more code that will make it so that if they leave the first name blank it will prompt asking if it's Mr., Mrs. or Ms. from there it will add Mr., Ms., Mrs. To the last name. 

In the event that neither is entered then it will default to friend.